+++ src/index-web.ts
@@ -10,6 +10,7 @@ export * from './chain/tezos/TezosMessageUtil';
 export * from './chain/tezos/TezosNodeReader';
 export * from './chain/tezos/TezosNodeWriter';
 export * from './chain/tezos/contracts/BabylonDelegationHelper';
+export * from './chain/tezos/contracts/CryptonomicNameServiceHelper';
 export * from './chain/tezos/contracts/DexterTokenHelper';
 export * from './chain/tezos/contracts/Tzip7ReferenceTokenHelper';
 export * from './chain/tezos/contracts/TCFBakerRegistryHelper';